# specs/001-backend-foundation/contracts/todo-api-openapi.yaml
openapi: 3.0.3
info:
  title: Todo Backend API
  description: Secure API for managing user tasks with JWT authentication
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
  /api/tasks:
    get:
      summary: Retrieve all tasks for the authenticated user
      description: Returns a list of tasks that belong to the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      summary: Create a new task for the authenticated user
      description: Creates a new task associated with the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/tasks/{id}:
    get:
      summary: Retrieve a specific task
      description: Returns a specific task that belongs to the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: A single task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
    put:
      summary: Update a specific task
      description: Updates a specific task that belongs to the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      summary: Delete a specific task
      description: Deletes a specific task that belongs to the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          example: "Milk, eggs, bread, fruits"
        completed:
          type: boolean
          example: false
        user_id:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          example: "Milk, eggs, bread, fruits"
        completed:
          type: boolean
          example: false
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          example: "Milk, eggs, bread, fruits"
        completed:
          type: boolean
          example: true
  responses:
    UnauthorizedError:
      description: Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Not authenticated"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Task not found"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Validation error"
  schemas:
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: "Error message"